refactor: split up codebase across reasonable files
