+++ jest.config.js
@@ -1,4 +1,9 @@
 module.exports = {
   preset: 'ts-jest',
-  testEnvironment: 'node'
+  testEnvironment: 'node',
+  testRegex: '(/__tests__/.*|(\\.|/)test)\\.ts$',
+  collectCoverageFrom: [
+    'src/**',
+    '!src/**/*.d.ts'
+  ]
 }
