+++ src/utils/__tests__/slack.test.ts
@@ -1,39 +1,6 @@
-import { getNowPlaying, shouldSetStatus } from '../utils'
+import { shouldSetStatus } from '../slack'
 
-describe('utils', () => {
-  describe('getNowPlaying', () => {
-    it('returns a now playing track', () => {
-      const track: Partial<LastFM.Track> = {
-        '@attr': {
-          nowplaying: 'true'
-        }
-      }
-
-      expect(getNowPlaying([track as any])).toEqual(track)
-    })
-
-    it('returns undefined when nowplaying = false', () => {
-      const track: Partial<LastFM.Track> = {
-        '@attr': {
-          nowplaying: 'false'
-        }
-      }
-
-      expect(getNowPlaying([track as any])).toBeUndefined()
-    })
-
-    it('returns undefined when @attr.nowplaying is undefined', () => {
-      const track: Partial<LastFM.Track> = {
-        artist: {
-          mbid: '',
-          '#text': ''
-        }
-      }
-
-      expect(getNowPlaying([track as any])).toBeUndefined()
-    })
-  })
-
+describe('slack', () => {
   describe('shouldSetStatus', () => {
     it('returns true when the app has previously updated the profile', () => {
       const profile: Partial<Slack.Profile> = {
